% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSMotifDiscover.R
\name{HSMotifDiscover}
\alias{HSMotifDiscover}
\title{HSMotifDiscover}
\usage{
HSMotifDiscover(
  input_HSseq_file,
  motifLenVec,
  charGrpFile,
  seq_weight_file,
  numCores = 1,
  affinity_threshold = 0,
  itr = 40000,
  if_revCompStrand = FALSE
)
}
\arguments{
\item{input_HSseq_file}{A Heparin Sulfate sequence (or any other seqeunce) file. The file contains header and sequence information of the samples similar to fasta file, but sequence character counts per line are not constrained.}

\item{motifLenVec}{Describe a vector of the motif lenghts to be discovered.}

\item{charGrpFile}{(Optional parameter) File having two columns , \emph{One}- dimers/trimers/tetramers that should be considered as single characters to discover motif,  \emph{Two}- character group information in numeric form. In heparin sulphate dimer and trimer occupy alternate positions are. So, dimers are grouped in one group and trimers are grouped in other group. If this file is not given then each character will be considered independently to discover motif as other motif discovery tools such as MEME, but will work for any other type of characters.}

\item{seq_weight_file}{(Optional parameter) File having column of sequence header and sequence weight in motif discovery. If this file is not selected then all sequences have equal weight in motif discovery.}

\item{numCores}{(Optional parameter) the number of cores to be used. This feature is useful when motif is discovered in large range. Multiple cores feature will not work on widows computer or RStudio environment.}

\item{affinity_threshold}{(Optional parameter) The sequences with weight greater than the threshold are used for motif discovery. This is required only of weight file is given.}

\item{itr}{(Optional parameter) Number of iterations for gibbs sampling optimisation. Higher itrations may improve the results but at the cost of time.}

\item{if_revCompStrand}{logical variable, TURE if reverse complementry strand should also be considered while motif discovery  (In case of DNA sequnces only). Deafault value is FALSE}
}
\value{
\enumerate{
\item \emph{list} A list of following parameters of all the discovered motifs in the given range.
\enumerate{
  \item  \emph{PSSM:} Position specific scoring matrix of the discovered motif.
  \item  \emph{MotifEntropy:} Average entropy of discovered motif.
  \item  \emph{IC:} Information content of the discovered motif.
  \item  \emph{bkgProb_vec:} Probability vector of the characters in the background sequences.
  \item  \emph{motif_Pval:} P-Value of the discovered motif.
  \item  \emph{CharDataf:} Summary of the input sequences characters.
  \item  \emph{orgChar_discoveredMotifs:} Sequences of the discovered motif.
  \item  \emph{MotifLogo:} Discovered motif seqlogo.
  \item  \emph{resultsTable_df:} Results of the discovered motif are summarised in the data frame. Where columns have following information
  \itemize{
     \item \emph{header:} header of the sequences used for motif discovery.
     \item \emph{sequence:} Input sequences.
     \item \emph{seqLen:} sequence length.
     \item \emph{if sequence weight is greater than the threshold:} boolean (TRUE/FALSE). This column is present only of weight of the sequences is given as input.
     \item \emph{max_score:} Maximum likelihood score of the of the PSSM in the sequnce.
     \item \emph{P_value:} The probability that a random sequence (with the same length and conforming to the background) would have position p-values such that the product is smaller or equal to the value calculated for the sequence under test.
     \item \emph{MotifLoc:} Start point of discovered motif. (This is the location of max_score in the sequence )
     \item \emph{seq_weights:} Input weights of the sequnces. This column is present only if weight of the sequences is given as input.
     \item \emph{score:} Liklihood score of the PSSM at different sequence locations.
     \item \emph{RC_strand:} '1' if motif is present on reverse complementry strand (Only in case of DNA sequences).
 }
 }
 \item \emph{MotifSummary_runTime_*.txt:} Summary of motifs in the text file. (* is the time of run)
 \item \emph{log_runTime_*.txt:} log file of the run. (* is the time of run)
 }
}
\description{
Discover motif in Heparan Sulphate (HS) sequences or any type of sequnces.
}
\details{
Discover motif in heparan sulphate sequences or any type of sequnces.
}
\examples{
HSSeq_file=system.file("extdata", "Ex_simulated_HSseqs_new_M1c_and_M3c_S100E.txt",
package = "HSMotifDiscover", mustWork = TRUE)
charGrpFile=system.file("extdata", "Chars.txt", package = "HSMotifDiscover", mustWork = TRUE)
seq_weight_file=system.file("extdata", "motif1c_Weight.txt", package = "HSMotifDiscover", mustWork = TRUE)
motifLenVec=c(5,7)   # motif length range

out=HSMotifDiscover(input_HSseq_file=HSSeq_file, motifLenVec=motifLenVec, charGrpFile=charGrpFile,
seq_weight_file=seq_weight_file, numCores=1,  affinity_threshold=0,  itr=40000)

###for DNA sequences ##
DNASeq_file=system.file("extdata", "SRF_syntheticSeqFile_Test.txt",
package = "HSMotifDiscover", mustWork = TRUE)
outDNA=HSMotifDiscover(input_HSseq_file=DNASeq_file, motifLenVec=10,
      numCores=1,  itr=50000, if_revCompStrand=TRUE)


}
